requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
openai-whisper==20231117
torch==2.1.0
transformers==4.35.0
numpy==1.24.3
pydantic==1.10.13
python-dotenv==1.0.0
aiofiles==23.2.1
sentencepiece==0.1.99
accelerate==0.21.0
scipy==1.10.1
protobuf==3.20.0
tokenizers>=0.13.3